patient-search/requirements.txt
fastapi
uvicorn
python-dotenv==0.19.1
farm-haystack[preprocessing,elasticsearch]